app/app_for_k8soperator/server/requirements.txt
blinker==1.7.0
click==8.1.7
Deprecated==1.2.14
Flask==2.3.3
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
opentelemetry-api==1.21.0
redis==5.0.1
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
